LayoutTests/http/tests/navigation/document-open-new-window-adds-history-item-expected.txt

This tests that document.open in a new window creates a back/forward item.

============== Back Forward List ==============
curr->  http://127.0.0.1:8000/navigation/document-open-new-window-adds-history-item.html  **nav target**
===============================================

============== Back Forward List ==============
curr->  about:blank  **nav target**
        http://127.0.0.1:8000/navigation/resources/document-open-page-2.html  **nav target**
===============================================
